hotfix for packaging issue
