+++ extensions/ezmsg-websocket/ezmsg/websocket/server.py
@@ -6,7 +6,21 @@ import websockets.exceptions
 
 import ezmsg.core as ez
 
-from . import WebsocketSettings, WebsocketState
+from dataclasses import field
+from typing import Optional, Union
+
+import ezmsg.core as ez
+
+class WebsocketSettings(ez.Settings):
+    host: str
+    port: int
+    cert_path: Optional[str] = None
+
+
+class WebsocketState(ez.State):
+    incoming_queue: "asyncio.Queue[Union[str,bytes]]" = field(default_factory=asyncio.Queue)
+    outgoing_queue: "asyncio.Queue[Union[str,bytes]]" = field(default_factory=asyncio.Queue)
+    
 
 class WebsocketServer(ez.Unit):
